src/faceExpression/requirements.txt
opencv-python==3.2.0.7
tensorflow-gpu==1.0.1
keras==2.0.3
